refactor: split web console command dispatch
